seperate game into components.
